# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Email Sender App
# during dev, should point to your local machine
host: localhost:5000
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  api_key:
    type: apiKey
    in: header
    name: api_key
paths:
  /attachment:
    x-swagger-router-controller: EmailController
    post:
      description: Test
      operationId: attachment
      parameters:
        - description: Attachment file
          in: formData
          name: file
          required: true
          type: file
      responses:
        "200":
          description: The attachment was uploaded
          schema:
            type: object
        "413":
            description: The attachment is too larger.
            schema:
              type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /send-mail:
    x-swagger-router-controller: hello_world
    post:
      description: Test
      operationId: sendmail
      parameters:
        - in: body
          name: maildata
          description: Data of mail
          schema:
            $ref: '#/definitions/EmailData'
      responses:
        "200":
          description: Mail was successfully sent
          schema:
            type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  EmailData:
    required:
      - subject
      - to
    properties:
      subject:
        type: string
      to:
        type: array
        items:
          type: string
      cc:
        type: array
        items:
          type: string
      text:
        type: string
      html:
        type: string
      attachmentIds:
          type: array
          items:
              type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string